Methods moved from Main to AnimalIO
